headlines_project/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
aiomysql==0.2.0
pydantic==2.11.7
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
